Advanced_Phase_Deliverables/Task_4_Anomaly_Detector/requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
langchain-google-genai
python-dotenv